Change gui and element access
All elements and the gui script are now global fields within the
editor_environment table, which is a separate environment specifically
for the editor.

The point of changing the environment is incase of wanting to embed the
editor inside of a love/lua game (which is something I want to do).

Making all those global just makes things more convenient