Create udimen
I need a better way to manage positioning and sizing of ui elements. At
some point I'd like to be able to not just position the objects based on
pixel units, but also position them relative to their parent. The same
way roblox has udims with scale and offset values